complete loader functionality
When user logs in, the dashboard layout will render.  But before it does, we invoke loader through useLoaderData hook to make another call for the info about the current user.  Currently, the only data we have is stored in the JWT/cookie, so we need to make another call to get all their data